BS_SCORE_backup/requirements.txt
Flask==2.0.1
numpy==1.20.3
pandas==1.2.4
scikit-learn==0.24.2
mysqlclient==2.0.3
mysql==0.0.3